requirements.txt
fastapi==0.59.0  #Framework Web
uvicorn==0.11.5  #Servidor
selenium==3.141.0
mysql.connector
spacy